Agregando CreateTail
Add a function called createTail
It should take the following parameters: head, dir, numCells
The head parameter is a reference to the randomly selected cell (within
createSnake)
The direction parameter is a normalized value on the X or Y axis (values
can be -1 or 1). Ideally this should be an object (use {} notation), but
if you're not comfortable with this notation use an array of two
elements were position 0 represents the X component and 1 represents the
Y component.
The numCells parameter represents the length of the snake's body. It
determines how many cells in a given direction are going to be active
besides the head.

Call the createTail function from within the createSnake function.
The expected output should be N amount of active cells within a given
direction.